+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Changed return value of low-level `parse` API to allow providing
   more detailed information about the TeX functions encountered.
 * Switch to text mode before emitting \AA and \textvisiblespace.
+* Require mandatory argument for \overbrace, \overleftarrow,
+  \overleftrightarrow, \overline, \overrightarrow, \underbrace, and
+  \underline.
 
 # texvcjs 0.2.0 (2014-07-23)